feat: move blog utils
