+++ apps/the_monkeys/src/app/blog/[slug]/utils.ts
